Split staticfiles tests in multiple files
Improve staticfiles test files structure.
